회원정보기능 / 회원탈퇴 기능 / 보안강화 구현완료

+++ board/views.py
@@ -3,11 +3,10 @@ from board.models import Post
 from django.core.paginator import Paginator
 from datetime import timezone
 from board.forms import PostCreate_Form
+from django.http import HttpResponseForbidden, HttpResponse
 
 
 def board_list(request):
-    
-
     page = request.GET.get("page", 1)
     post_list = Post.objects.filter(is_deleted=False, display_avilable=False).select_related("author")
     page_count = Paginator(post_list , 8)
@@ -24,33 +23,38 @@ def board_list(request):
 
 
 def board_create(request):
-    if request.method == "POST":
-        form = PostCreate_Form(request.POST, request.FILES)
-        if form.is_valid():
-            post = form.save(commit=False)
-            post.author = request.user
-            post.slug = form.cleaned_data.get("title")
-            post.save()
-            return redirect(Post.get_absolute_url(post))
+    if request.user.is_deleted == False:
+        if request.method == "POST":
+            form = PostCreate_Form(request.POST, request.FILES)
+            if form.is_valid():
+                post = form.save(commit=False)
+                post.author = request.user
+                post.slug = form.cleaned_data.get("title")
+                post.save()
+                return redirect(Post.get_absolute_url(post))
+        else:
+            form = PostCreate_Form()
+        return render(request, "board/board_create.html", {"form": form})
     else:
-        form = PostCreate_Form()
-    return render(request, "board/board_create.html", {"form": form})
+        return HttpResponseForbidden
 
 
 def board_update(request, pk, slug):
     post = get_object_or_404(Post, id=pk, slug=slug)
-    if request.method == "POST":
-        form = PostCreate_Form(request.POST, request.FILES, instance=post)
-        if form.is_valid():
-            post = form.save(commit=False)
-            post.author = request.user
-            post.slug = form.cleaned_data.get("title")
-            post.save()
-            return redirect(Post.get_absolute_url(post))
+    if request.user == post.author:
+        if request.method == "POST":
+            form = PostCreate_Form(request.POST, request.FILES, instance=post)
+            if form.is_valid():
+                post = form.save(commit=False)
+                post.author = request.user
+                post.slug = form.cleaned_data.get("title")
+                post.save()
+                return redirect(Post.get_absolute_url(post))
+        else:
+            form = PostCreate_Form(instance=post)
+        return render(request, "board/board_create.html", {"form": form})
     else:
-        form = PostCreate_Form(instance=post)
-    return render(request, "board/board_create.html", {"form": form})
-
+        return HttpResponseForbidden()
 
 
 def board_detail(request, pk, slug):
@@ -60,8 +64,20 @@ def board_detail(request, pk, slug):
 
 def board_delete(request, pk, slug):
     post = get_object_or_404(Post, id=pk, slug=slug)
-    post.delete()
-    return redirect("board:board_list")
+    if request.user == post.author:
+        post.delete()
+        return redirect("board:board_list")
+    else:
+        return HttpResponseForbidden()
 
 
 
+def board_likes(request):
+    if request.is_ajax():
+        post_id = request.GET.get("post_id")
+        post = Post.objects.get(id=post_id)
+
+        if not request.user.is_authenticated:
+            message = "로그인을 해주세욥"
+            context = {"like_count": post.like.count(), "message": message}
+            return http
